requirements.txt
flask
flask_cors
db-sqlite3
bcrypt
requests
google-cloud-vision
dateutils